Problem Exists Between Chair and Keyboard.
Somehow some files got misplaced and deleted, the OCD in me empties the trash alll the time.

Was in the middle of implementing auth with firebase when decided to just do it on my own project instead of doing it twice, since the login method needs to be tweeked a bit to meet the apps requirments of only user name and firebase's need for email and password.